# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Scootware Forum API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shoutbox:
    get:
      operationId: getShoutboxMessages
      tags: [shoutbox]
      summary: Get recent shoutbox messages
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Shoutbox messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShoutboxMessage"
    post:
      operationId: postShoutboxMessage
      tags: [shoutbox]
      summary: Post a shoutbox message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShoutboxMessageRequest"
      responses:
        "201":
          description: Message posted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShoutboxMessage"
  /shoutbox/{id}:
    delete:
      operationId: deleteShoutboxMessage
      tags: [shoutbox]
      summary: Delete a shoutbox message (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Message deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Message not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    ShoutboxMessage:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        authorId:
          type: integer
        authorUsername:
          type: string
        authorAvatarUrl:
          type: string
          nullable: true
        authorRole:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, content, authorId, authorUsername, authorRole, createdAt]
    CreateShoutboxMessageRequest:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 500
      required: [content]
...